Verify installation

+++ scripts/linux-setup.sh
@@ -12,8 +12,12 @@ sudo apt-get update
 p_success "Package lists updated."
 
 p_title "CURL"
-sudo apt -y install curl
-p_success "CURL installed."
+if cmd_exists curl ; then
+  p_warn "curl already installed."
+else
+  sudo apt -y install curl
+  p_success "CURL installed."
+fi
 
 p_title "Essentials"
 sudo apt-get -y install build-essential
