+++ scripts/neovim.sh
@@ -3,8 +3,12 @@
 source ./scripts/helpers.sh;
 
 p_title "Neovim"
-brew install neovim
-p_success "Neovim installed."
+if cmd_exists nvim ; then
+    p_warn "Neovim already installed."
+else
+    brew install neovim
+    p_success "Neovim installed."
+fi
 
 p_title "Vim-plug"
 sh -c 'curl -fLo "${XDG_DATA_HOME:-$HOME/.local/share}"/nvim/site/autoload/plug.vim --create-dirs \
@@ -14,8 +18,8 @@ p_success "Vim-plug"
 p_title "Neovim configuration files"
 NEOVIM_PATH=${HOME}/.config/nvim
 mkdir -p $NEOVIM_PATH
-cp ./configs/neovim ${NEOVIM_PATH}/init.vim
-cp ./configs/bundles.vim ${NEOVIM_PATH}/bundles.vim
+cp ./config-files/neovim ${NEOVIM_PATH}/init.vim
+cp ./config-files/bundles.vim ${NEOVIM_PATH}/bundles.vim
 touch ${NEOVIM_PATH}/bundle.vim.local
 p_success "Neovim configuration files"
